Moved ImageUtil into a more neutral place, and added an explicit tool for histogram comparison.
